Enum small_step_simple::Element

pub enum Element {
    Number(i64),
    Add(Box<Element>, Box<Element>),
    Multiply(Box<Element>, Box<Element>),
    Boolean(bool),
    LessThan(Box<Element>, Box<Element>),
    Variable(String),
    Assign(String, Box<Element>),
    Sequence(Box<Element>, Box<Element>),
    IfElse(Box<Element>, Box<Element>, Box<Element>),
    While(Box<Element>, Box<Element>),
    DoNothing,
}

Our AST elements.

Variants

Number

A simple number object, this cannot be reduced further.

Add

An addition of two elements.

Multiply

A multiplication of two elements.

Boolean

A simple boolean object, this cannot be reduced further.

LessThan

A less-than relation check of two elements. Elements should reduce to a number to be comparable.

Variable

A variable, will be replaced by its value when reducing.

Assign

A variable assignment. Only completely reduced values are assigned. No type checks.

Sequence

A sequence of two elements. The first element is reduced completely before the second is touched.

IfElse

A if-else block. Condition needs to reduce to a Boolean. No type checking. If condition reduces to true, the consequence is used furhter, otherwise the alternative

While

A while loop. Runs until the condition reduces to false.

DoNothing

A simple no-op statement.

Methods

impl Element

fn is_reducible(&self) -> bool

Wether or not an expression is reducible. See Element for more info.

fn value(&self) -> i64

Get the actual value of a Number. Fails for other elements than Number and Boolean. Boolean maps to Integers: true=1, false=0.

fn reduce(&self, environment: &mut HashMap<String, Box<Element>>) -> Element

Reduce the expression according to the rules for the current element.