Struct small_step_simple::Machine

pub struct Machine {
    // some fields omitted
}

Our virtual machine, executing our constructed AST step-by-step

Methods

impl Machine

fn new(expression: Box<Element>, map: HashMap<String, Box<Element>>) -> Machine

Create a new machine with a given expression and an environment

fn new_with_empty_env(expression: Box<Element>) -> Machine

Create a new machine with a given expression and an empty environment

fn clone_env(&self) -> HashMap<String, Box<Element>>

As the environment is passed in immutable, we need to clone it to get it back

fn step(&mut self)

Reduce one step of our current expression

fn run(&mut self)

Reduce until we reached a non-reducible expression. This prints the current expression before each step.